requirements.txt
PyYAML>=3.11
requests>=2.4.0
-e git+https://git.openstack.org/openstack-infra/zuul#egg=zuul